Make hooks only apply to staged files
Improve workflow as follows:
* Refine hooks by only linting staged .{js,jsx,ts,tsx} files. If there
are any such staged files, run tsc --noEmit on the whole project to
check for any type errors. This is to increase accuracy of
linting/testing.
* Only apply commit/push hooks to master and preview branches. This is
to encourage the usage of other branches for backup/contributing
purposes.

This commit introduces two main dependencies: git-branch-is (for
portability purposes; we want the scripts to be runnable on both Windows
and POSIX shells) and lint-staged.